Update README.md
Clarified UML requirements (CLASS Diagram, incorporate Strategy and Factory Design Patterns and specify their relationships to the Arena and Braitenberg Vehicles Classes.   Noted that we refer to robots and Braitenberg vehicles interchangeably.  Added more to specification of the Factory and Strategy Pattern - and relaxed the requirement for the Factory pattern.